pyproject.toml
[project]
name = "morvex"
version = "0.1.0"
description = ""
authors = [
    { name = "Nima Nooshiri", email = "nima.nooshiri@gmail.com" }
]
readme = "README.md"
license = 'MIT'
license-files = ['LICENSE']
requires-python = ">=3.12"
dependencies = [
    "cmcrameri>=1.9",
    "cupy-cuda11x>=13.6.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "rich>=14.1.0",
    "scipy>=1.16.2",
]

[project.optional-dependencies]
plotly = [
    "plotly>=6.3.0",
]
marimo = [
    "marimo>=0.15.5",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipython>=9.5.0",
    "pre-commit>=4.3.0",
    "prek>=0.3.3",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
    "ty>=0.0.1a20",
]

[tool.ty.rules]
unresolved-attribute = "ignore"
